major refactoring to simplify the code setup: - move 'kernels' interface functions to 'mathfuncs.py', - only provde one backend, namely cuda_kernels.py, which are   launched directly from mathfuncs.py
THERE IS NO MORE CPU IMPLEMENTATION
